experimentally use float8 in leaves of 'spoint2' indexes (renamed to 'spoint3')
Storing the full precision of Spoint data in the leaves of the
'spoint2' operator class appears to be an adequate means to
make OPERATOR=(spoint, spoint) work for corresponding indexes,
although this will increase the index size by 8 to 12 bytes
for each indexed row.

The operator class 'spoint2' was pragmatically renamed to
'spoint3' in order to prevent chaos and crashes in existing
PgSphere installations.